int battery in Scooter model

+++ src/main/java/com/softserve/kickscooter/kickscootervehicle/management/model/Scooter.java
@@ -33,7 +33,7 @@ public class Scooter {
 
     private double actualLatitude;
 
-    private short battery;
+    private int battery;
 
 
 }
